+++ src/main/java/com/cinchapi/runway/access/Scope.java
@@ -17,14 +17,11 @@ package com.cinchapi.runway.access;
 
 import javax.annotation.concurrent.Immutable;
 
-import com.cinchapi.ccl.syntax.ConditionTree;
 import com.cinchapi.common.reflect.Reflection;
-import com.cinchapi.concourse.lang.ConcourseCompiler;
 import com.cinchapi.concourse.lang.Criteria;
 import com.cinchapi.runway.Record;
 import com.cinchapi.runway.Selection;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Multimap;
 
 /**
  * Describes the visibility that an {@link Audience} has for a given
@@ -163,10 +160,7 @@ public abstract class Scope {
 
         @Override
         public boolean test(Record record) {
-            ConcourseCompiler compiler = ConcourseCompiler.get();
-            ConditionTree tree = (ConditionTree) compiler.parse(criteria);
-            Multimap<String, Object> mmap = Reflection.call(record, "mmap");
-            return compiler.evaluate(tree, mmap); /* (authorized) */
+            return record.matches(criteria);
         }
     }